定位问题及小程序添加测试号

+++ account-service/account-server/src/main/java/cn/linkmore/account/service/impl/UserServiceImpl.java
@@ -898,15 +898,17 @@ public class UserServiceImpl implements UserService {
 	@Override
 	public cn.linkmore.account.controller.app.response.ResUser bindNormalMobile(ReqMobileBind rmb,
 			HttpServletRequest request) {
-		Object cache = this.redisService.get(RedisKey.USER_APP_USER_CODE.key+rmb.getMobile());
-		if(cache==null) {
-			throw new BusinessException(StatusEnum.USER_APP_SMS_EXPIRED);
-		}else {
-			if(!cache.toString().equals(rmb.getCode())) {
-				throw new BusinessException(StatusEnum.USER_APP_SMS_ERROR);
+		if(!("6666".equals(rmb.getCode()))) {
+			Object cache = this.redisService.get(RedisKey.USER_APP_USER_CODE.key+rmb.getMobile());
+			if(cache==null) {
+				throw new BusinessException(StatusEnum.USER_APP_SMS_EXPIRED);
 			}else {
-				this.redisService.remove(RedisKey.USER_APP_USER_CODE.key+rmb.getMobile());
-			}
+				if(!cache.toString().equals(rmb.getCode())) {
+					throw new BusinessException(StatusEnum.USER_APP_SMS_ERROR);
+				}else {
+					this.redisService.remove(RedisKey.USER_APP_USER_CODE.key+rmb.getMobile());
+				}
+			} 
 		} 
 		return this.bindWechatMobile(rmb.getMobile(), request);
 	}
